Split source code for backend into multiple files.
